requirements.txt
xlrd==1.2.0
pandas==0.25.3
requests==2.25.1
unidecode==1.3.3
psycopg2-binary==2.8.6
protobuf==3.19.4
mysql-connector-python==8.0.27
google-api-python-client==2.48.0
oauth2client==4.1.3